requirements.txt
Django==4.1.*
djangorestframework==3.14.*
Pillow==9.5.*
celery==5.2.*
redis==4.3.*
django-constance==4.3.*
django-leaflet==0.28.*
openpyxl==3.0.*
xlsxwriter==3.0.*
django-summernote==0.8.*
psycopg2-binary==2.9.*
django-celery-beat==2.7.0